[MT] CIrrDeviceSDL: Fix numpad key events not having correct KeyInput.Char
Backported by SmallJoker. Origin: Minetest. Hash: 9a1501ae8, Date: 2024-06-27

Allows you to change viewing range using numpad +/- again. This fix also works
with the current unreleased version of SDL 3.

The keycodes for numpad keys (both SDL keycodes and Irrlicht keycodes) are not
the same as the keycodes for the equivalent non-numpad keys and don't
correspond to chars, so I mapped them to chars manually.

Since I think the resolution of minetest/minetest/issues/13770 was
"just disable numlock", I made sure to only do this for the numpad number keys
if numlock is enabled.