[MT] Refactor matrix4.h
Backported by SmallJoker. Origin: Minetest. Hash: 7e4919c6e, Date: 2024-10-10

Sets the surprising row-major conventions used here straight.

Renames rotateVect to rotateAndScaleVect:
If the matrix also scales, that is applied as well by the method.
Obsolete rotateVect variants are removed.
The inverseRotateVect method is also renamed accordingly.
Note that this applies the transpose of the product
of the scale and rotation matrices, which inverts just the rotation.