[LT] Fix operator[] for vector2d and vector3d being potentially UB
Backported commit. Origin: Luanti. Hash: a00b9cab3, Date: 2025-04-08

We don't have a C++ expert on hand, but taking a pointer to one member
and expecting to access another by an offset is very fishy:
- for one, there could theoretically be padding
- the compiler might assume that we are only writing to that first member

The new code has shown to be free for constant parameter values.
Non-constant ones cause the assembly to have branches (why?), but we don't
use that much.